unit test Packet : Ping Pong Ack Event DeleteEntity ErrorGame ErrorHandshake ErrorList ErrorScore EventError gameList join

+++ core/src/network/packet/PacketDeleteEntity.cpp
@@ -33,7 +33,7 @@ uint32_t network::packet::PacketDeleteEntity::getTick() const
     return GET_NETINT(schema, 0);
 }
 
-void network::packet::PacketDeleteEntity::setEventId(uint16_t id)
+void network::packet::PacketDeleteEntity::setEventId(uint32_t id)
 {
     SET_NETINT(schema, 1, id);
 }
